Controller System Added!
Controller.java has components that should act as an appropriate movement system for the project that should be easily implemented. When pressing the start button, there is a black square which indicates which movement option you picked last and also restricts vertical/horizontal directions depending on the direction you went last.